services/model-router/requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
pydantic>=2.3.0
pydantic-settings>=2.0.3
sqlalchemy>=2.0.21
aiohttp>=3.8.5
asyncio>=3.4.3
prometheus-client>=0.17.1
tiktoken>=0.5.1
pillow>=10.0.1
python-jose>=3.3.0
python-multipart>=0.0.6
tenacity>=8.2.3
pytest>=7.4.2
httpx>=0.25.0
pytest-asyncio>=0.21.1